added table headers
